+++ libs/textile-js/textile.js
@@ -1,4 +1,4 @@
-// [textile-js | https://github.com/GehDoc/textile-js]  Build version: 2.0.106 - 2020-03-29T17:59:02+0200  
+// [textile-js | https://github.com/GehDoc/textile-js]  Build version: 2.0.106 - 2020-03-29T23:25:02+0200  
  (function webpackUniversalModuleDefinition(root, factory) {
 	if(typeof exports === 'object' && typeof module === 'object')
 		module.exports = factory();
@@ -1081,7 +1081,7 @@ function escape(text, escapeQuotes) {
   return text.replace(/&(?!(#\d{2,}|#x[\da-fA-F]{2,}|[a-zA-Z][a-zA-Z1-4]{1,6});)/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, escapeQuotes ? '&quot;' : '"').replace(/'/g, escapeQuotes ? '&#39;' : "'");
 }
 
-function toHTML(jsonml) {
+function toHTML(jsonml, render) {
   jsonml = jsonml.concat(); // basic case
 
   if (typeof jsonml === 'string') {
@@ -1098,20 +1098,21 @@ function toHTML(jsonml) {
   }
 
   while (jsonml.length) {
-    content.push(toHTML(jsonml.shift()));
+    content.push(toHTML(jsonml.shift(), render));
   }
 
   for (var a in attributes) {
     tagAttrs += attributes[a] == null ? " ".concat(a) : " ".concat(a, "=\"").concat(escape(String(attributes[a]), true), "\"");
-  } // be careful about adding whitespace here for inline elements
+  }
 
+  var realContent = render && render.content ? render.content(tag, attributes, content.join('')) : content.join(''); // be careful about adding whitespace here for inline elements
 
   if (tag === '!') {
-    return "<!--".concat(content.join(''), "-->");
+    return "<!--".concat(realContent, "-->");
   } else if (tag in singletons || tag.indexOf(':') > -1 && !content.length) {
     return "<".concat(tag).concat(tagAttrs, " />");
   } else {
-    return "<".concat(tag).concat(tagAttrs, ">").concat(content.join(''), "</").concat(tag, ">");
+    return "<".concat(tag).concat(tagAttrs, ">").concat(realContent, "</").concat(tag, ">");
   }
 }
 
@@ -1733,7 +1734,9 @@ function textile(txt, opt) {
   // get a throw-away copy of options
   opt = merge(merge({}, textile.defaults), opt || {}); // run the converter
 
-  return parseFlow(txt, opt, opt.lineOffset).map(jsonmlUtils.toHTML).join('');
+  return parseFlow(txt, opt, opt.lineOffset).map(function (value) {
+    return jsonmlUtils.toHTML(value, opt.render);
+  }).join('');
 }
 
 ;
@@ -1749,7 +1752,9 @@ textile.defaults = {
   // by default, don't set a special CSS class name to each HTML element mapped to an original line number
   'cssClassOriginalLineNumber': '',
   // functions to apply to each JsonML node before rendering to HTML
-  'hooks': []
+  'hooks': [],
+  // function called where a JsonML node is rendered to HTML
+  'render': {}
 };
 
 textile.setOptions = textile.setoptions = function (opt) {
@@ -1772,7 +1777,13 @@ textile.jsonml = function (txt, opt) {
   return ['html'].concat(textile.tokenize(txt, opt));
 };
 
-textile.serialize = jsonmlUtils.toHTML;
+textile.serialize = function (jsonml, opt) {
+  // get a throw-away copy of options
+  opt = merge(merge({}, textile.defaults), opt || {}); // serialize
+
+  return jsonmlUtils.toHTML(jsonml, opt.render);
+};
+
 textile.jsonmlUtils = jsonmlUtils;
 
 /***/ }),